Fix compiler warnings. Fixes bug #322053. Patch provided by Frederic
2005-11-22  Brian Cameron <Brian.Cameron@sun.com>

        * daemon/cookie.c, daemon/gdm.c, daemon/xdmcp.c,
          gui/gdmchooser.c, gui/gdmlogin.c,
          gui/gdmphotosetup.c, gui/gdmsetup.c,
          gui/gdmwm.c, gui/greeter/greeter_events.c,
          gui/greeter/greeter_item_ulist.c,
          gui/greeter/greeter_parser.[ch],
          gui/modules/dwellmouselistener.c,
          gui/modules/keymouselistener.c:  Fix
          compiler warnings.  Fixes bug #322053.
          Patch provided by Frederic Crozat.

+++ gui/greeter/greeter_events.c
@@ -84,9 +84,9 @@ propagate_state (GreeterItemInfo *info,
   state_run (info, old_state);
 
   g_list_foreach (info->fixed_children, propagate_state_foreach,
-		  GINT_TO_POINTER (info->state));
+		  GINT_TO_POINTER ((int) info->state));
   g_list_foreach (info->box_children, propagate_state_foreach,
-		  GINT_TO_POINTER (info->state));
+		  GINT_TO_POINTER ((int) info->state));
 }
 
 static void propagate_reset_state (GreeterItemInfo *info,
